requirements.txt
transformers==4.48.0
tokenizers==0.21.0
huggingface-hub==0.26.2

fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.4
pydantic-settings==2.4.0
python-dotenv==1.0.1
python-multipart==0.0.9

numpy==1.26.4

datasets==2.18.0
sentencepiece==0.2.0
tqdm==4.66.2
protobuf==4.25.3

langchain==0.3.15
langchain-core==0.3.35
langchain-community==0.3.15
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6

faiss-cpu==1.8.0
pypdf==4.2.0

sentence-transformers==3.0.1
unstructured==0.14.6
